main_test.go
package main

import (
	"bytes"
	"net/http"
	"net/http/httptest"
	"os"
	"testing"
)

// Define the API type
type API struct{}

// Define the GetReceiptByID method
func (api *API) GetReceiptByID(w http.ResponseWriter, r *http.Request) {
	// Placeholder implementation
}

// Define the CreateReceipt method
func (api *API) CreateReceipt(w http.ResponseWriter, r *http.Request) {
	// Placeholder implementation
}

func TestGetReceiptByID(t *testing.T) {
	// Create a new instance of the API
	api := &API{}

	// Create a new request
	req, err := http.NewRequest("GET", "/receipt/123", nil)
	if err != nil {
		t.Fatal(err)
	}

	// Create a new response recorder to capture the response
	rr := httptest.NewRecorder()

	// Call the GetReceiptByID method
	api.GetReceiptByID(rr, req)

	// Check the response status code
	if rr.Code != http.StatusOK {
		t.Errorf("expected status %d but got %d", http.StatusOK, rr.Code)
	}

	// Check the response body
	expectedResponse := `{"Total Points":0,"Breakdown":""}` // Update with the expected response
	if rr.Body.String() != expectedResponse {
		t.Errorf("expected response %s but got %s", expectedResponse, rr.Body.String())
	}
}

func TestCreateReceipt(t *testing.T) {
	// Create a new instance of the API
	api := &API{}

	// Create a sample request body
	requestBody := []byte(`{
		"retailer": "Target",
		"purchaseDate": "2022-01-01",
		"purchaseTime": "13:01",
		"items": [
			{
				"shortDescription": "Mountain Dew 12PK",
				"price": 6.49
			},
			{
				"shortDescription": "Emils Cheese Pizza",
				"price": 12.25
			},
			{
				"shortDescription": "Knorr Creamy Chicken",
				"price": 1.26
			},
			{
				"shortDescription": "Doritos Nacho Cheese",
				"price": 3.35
			},
			{
				"shortDescription": "Klarbrunn 12-PK 12 FL OZ",
				"price": 12.00
			}
		],
		"total": 35.35
	}`)

	// Create a new request
	req, err := http.NewRequest("POST", "/receipt", bytes.NewBuffer(requestBody))
	if err != nil {
		t.Fatal(err)
	}

	// Create a new response recorder to capture the response
	rr := httptest.NewRecorder()

	// Call the CreateReceipt method
	api.CreateReceipt(rr, req)

	// Check the response status code
	if rr.Code != http.StatusOK {
		t.Errorf("expected status %d but got %d", http.StatusOK, rr.Code)
	}

	// Check the response body
	expectedResponse := `{"message":"Receipt created successfully"}` // Update with the expected response
	if rr.Body.String() != expectedResponse {
		t.Errorf("expected response %s but got %s", expectedResponse, rr.Body.String())
	}
}

func TestMain(m *testing.M) {
	// Run tests
	code := m.Run()

	// Exit with the test result code
	os.Exit(code)
}
